% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_season.R
\name{date_season}
\alias{date_season}
\alias{date_season.default}
\alias{date_season.seasons}
\title{Get the begin and end dates of seasons in seasonal time series}
\usage{
date_season(x, ...)

\method{date_season}{default}(x, ...)

\method{date_season}{seasons}(x, ...)
}
\arguments{
\item{x}{A numeric vector or an object of class seasons}

\item{...}{further argument to be passed to the class constructor function \code{seasons ()}.
Ignored when \code{x} is of class \code{seasons}.}
}
\value{
The beginning and the end of the season.
}
\description{
Getting the begin and end dates of seasons in seasonal time series such as rainfall data.
When the rainy season is bimodal, for instance, the function detect the 2 seasons first, then
return the begin and end dates of each.
}
\details{
detecting the dates of the biginning and the end of the season is more likely to
yield meaningful result when \code{x} has been derived via \code{get_phases (x,..., returned = 'seasonal')} which
rescales the data to make more consistant ordering.
}
\examples{
nam<- seq.Date(from = as.Date('2016-01-01'), to = as.Date ('2018-12-31'), by = 16)
dy11 <- c(1.40, 1.00, 1.50, 2.00, 5.00, 3.00, 1.00, 0.76, 2.00, 1.00, 3.50, 3.00, 1.50)
dy12 <- c(1.30, 1.10, 1.40, 2.01, 5.50, 2.80, 1.01, 1, 2.03, 1.09, 3.10, 3.00, 1.50)
dy1 <- c(dy11, dy12)
names(dy1) <- nam[1:length(dy1)]

y <- get_phases (dy1, ts_freq = 12)
z <- lapply(y, date_season)
}
\seealso{
\code{\link[growingSeason]{phases}}, \code{\link[growingSeason]{get_phases}}
}
\author{
Issoufou Liman

Issoufou Liman
}
