% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_seasonal_ranges.R
\name{ggplot_seasonal_ranges}
\alias{ggplot_seasonal_ranges}
\alias{ggplot_seasonal_ranges.default}
\alias{ggplot_seasonal_ranges.seasonal_ranges}
\alias{ggplot_seasonal_ranges.list}
\title{ggplot seasonal ranges}
\usage{
ggplot_seasonal_ranges(x, show_limits = TRUE, ...,
  outlier.colour = "black", outlier.color = "black",
  outlier.fill = "black", outlier.shape = 19, outlier.size = 1e-05,
  outlier.stroke = 0.5, outlier.alpha = 0.01, na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE, lwd = 0.1)

\method{ggplot_seasonal_ranges}{default}(x, show_limits = TRUE, ...,
  outlier.colour = "black", outlier.color = "black",
  outlier.fill = "black", outlier.shape = 19, outlier.size = 1e-05,
  outlier.stroke = 0.5, outlier.alpha = 0.01, na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE, lwd = 0.1)

\method{ggplot_seasonal_ranges}{seasonal_ranges}(x, show_limits = TRUE,
  ..., outlier.colour = "black", outlier.color = "black",
  outlier.fill = "black", outlier.shape = 19, outlier.size = 1e-05,
  outlier.stroke = 0.5, outlier.alpha = 0.01, na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE, lwd = 0.1)

\method{ggplot_seasonal_ranges}{list}(x, show_limits = TRUE, ...,
  outlier.colour = "black", outlier.color = "black",
  outlier.fill = "black", outlier.shape = 19, outlier.size = 1e-05,
  outlier.stroke = 0.5, outlier.alpha = 0.01, na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE, lwd = 0.1)
}
\arguments{
\item{x}{an object of class dataframe, matrix, or RasterStack. Names of x should
be dates in the format \code{X\%Y.\%m.\%d}.}

\item{show_limits}{Logical.Should the differents ranges be delineated with lines?}

\item{...}{Additional arguments passed to seasonal_ranges. This is not used with \code{ggplot_seasonal_ranges.seasonal_ranges}.}

\item{outlier.colour}{Default aesthetics for outliers. Set to \code{NULL} to inherit from the
aesthetics used for the box.

In the unlikely event you specify both US and UK spellings of colour, the
US spelling will take precedence.

Sometimes it can be useful to hide the outliers, for example when overlaying
the raw data points on top of the boxplot. Hiding the outliers can be achieved
by setting \code{outlier.shape = NA}. Importantly, this does not remove the outliers,
it only hides them, so the range calculated for the y-axis will be the
same with outliers shown and outliers hidden.}

\item{outlier.color}{Default aesthetics for outliers. Set to \code{NULL} to inherit from the
aesthetics used for the box.

In the unlikely event you specify both US and UK spellings of colour, the
US spelling will take precedence.

Sometimes it can be useful to hide the outliers, for example when overlaying
the raw data points on top of the boxplot. Hiding the outliers can be achieved
by setting \code{outlier.shape = NA}. Importantly, this does not remove the outliers,
it only hides them, so the range calculated for the y-axis will be the
same with outliers shown and outliers hidden.}

\item{outlier.fill}{Default aesthetics for outliers. Set to \code{NULL} to inherit from the
aesthetics used for the box.

In the unlikely event you specify both US and UK spellings of colour, the
US spelling will take precedence.

Sometimes it can be useful to hide the outliers, for example when overlaying
the raw data points on top of the boxplot. Hiding the outliers can be achieved
by setting \code{outlier.shape = NA}. Importantly, this does not remove the outliers,
it only hides them, so the range calculated for the y-axis will be the
same with outliers shown and outliers hidden.}

\item{outlier.shape}{Default aesthetics for outliers. Set to \code{NULL} to inherit from the
aesthetics used for the box.

In the unlikely event you specify both US and UK spellings of colour, the
US spelling will take precedence.

Sometimes it can be useful to hide the outliers, for example when overlaying
the raw data points on top of the boxplot. Hiding the outliers can be achieved
by setting \code{outlier.shape = NA}. Importantly, this does not remove the outliers,
it only hides them, so the range calculated for the y-axis will be the
same with outliers shown and outliers hidden.}

\item{outlier.size}{Default aesthetics for outliers. Set to \code{NULL} to inherit from the
aesthetics used for the box.

In the unlikely event you specify both US and UK spellings of colour, the
US spelling will take precedence.

Sometimes it can be useful to hide the outliers, for example when overlaying
the raw data points on top of the boxplot. Hiding the outliers can be achieved
by setting \code{outlier.shape = NA}. Importantly, this does not remove the outliers,
it only hides them, so the range calculated for the y-axis will be the
same with outliers shown and outliers hidden.}

\item{outlier.stroke}{Default aesthetics for outliers. Set to \code{NULL} to inherit from the
aesthetics used for the box.

In the unlikely event you specify both US and UK spellings of colour, the
US spelling will take precedence.

Sometimes it can be useful to hide the outliers, for example when overlaying
the raw data points on top of the boxplot. Hiding the outliers can be achieved
by setting \code{outlier.shape = NA}. Importantly, this does not remove the outliers,
it only hides them, so the range calculated for the y-axis will be the
same with outliers shown and outliers hidden.}

\item{outlier.alpha}{Default aesthetics for outliers. Set to \code{NULL} to inherit from the
aesthetics used for the box.

In the unlikely event you specify both US and UK spellings of colour, the
US spelling will take precedence.

Sometimes it can be useful to hide the outliers, for example when overlaying
the raw data points on top of the boxplot. Hiding the outliers can be achieved
by setting \code{outlier.shape = NA}. Importantly, this does not remove the outliers,
it only hides them, so the range calculated for the y-axis will be the
same with outliers shown and outliers hidden.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[=borders]{borders()}}.}

\item{lwd}{Numeric. line width}
}
\description{
ggplot seasonal ranges
}
