% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_stack_season_dates.R
\name{par_stack_season_dates}
\alias{par_stack_season_dates}
\title{Estimate and stack the beginning and end of growing seasons using parallel processing.}
\usage{
par_stack_season_dates(
  x,
  type = "v_points",
  n_criticals = 1,
  steps = 2,
  ts_freq = 23,
  returned = c("ts_seasonal", "original"),
  valid_range = NULL,
  filename = "",
  ...
)
}
\arguments{
\item{x}{A Raster \code{\link[raster]{brick}} or \code{\link[raster]{stack}}}

\item{type}{character string. either \code{v_points} or \code{peaks} depending on how \code{x} should be
broken into phases. \code{v_points} (default) may make more sens for expression seasonal data.}

\item{n_criticals}{Numeric. The number of points to be considered. Default to 1.}

\item{steps}{Integer. The minimum number of points at either sides of each v-like point
to be considered higher for it to qualify as v-like point. Default to 3}

\item{ts_freq}{The frequence of time series (see frequency argument in \code{\link[stats]{ts}}). Default to 23.}

\item{returned}{character string either \code{ts_seasonal} or \code{original.} if \code{ts_seasonal} (default),
the returned phases data will be the seasonal component of the decomposed time series (see \code{\link[stats]{decompose}}).
if 'original', the original data will be returned.}

\item{valid_range}{A numeric vector of lenght 2. Data cleaning parameter specifying the validity domain of the input data are valid.
Values outside this range are set NA before any processing. In the case of normalized difference indices (e.g. NDVI, SAVI) for example,
one should set these values to be between.}

\item{filename}{Output filename}

\item{...}{Additional arguments: 
  		
\code{datatype}{Character. Output data type (e.g. 'INT2S' or 'FLT4S'). 
See \code{\link[raster]{dataType}}. If no datatype is specified, 'FLT4S' is used,
unless this default value was changed with \code{\link[raster]{rasterOptions}}}
	
\code{overwrite}: Logical. If TRUE, "filename" will be overwritten if it exists
	
\code{progress}: Character. Set a value to show a progress bar. Valid values are "text" and "window".

\code{NAflag}: Numeric. To overwrite the default value used to represent \code{NA} in a file 
	
\code{bandorder}: Character. 'BIL', 'BIP', or 'BSQ'. For 'native' file formats only.
For some other formats you can use the 'options' argument (see below)
	
\code{options}: Character. File format specific GDAL options. E.g., when
writing a geotiff file you can use: \code{options=c("COMPRESS=NONE", "TFW=YES")}

You can use options=c("PROFILE=BASELINE") to create a plain  tif with no GeoTIFF tags. 
This can be useful when writing files to be read by applications intolerant of unrecognised tags. 
(see \url{http://www.gdal.org/frmt_gtiff.html})

NetCDF files have the following additional, optional, arguments: \code{varname}, \code{varunit}, \code{longname}, \code{xname}, \code{yname}, \code{zname}, \code{zunit}

\code{prj}: Logical. If \code{TRUE}, the crs is written to a .prj file. This can be useful 
when writing to an ascii file or another file type that does not store the crs
}
}
\value{
A \code{\link[raster]{stack}} object of biginning and end dates of the season for each pixel.
}
\description{
Estimate and stack the beginning and end of growing seasons using parallel processing.
}
\seealso{
\code{\link[growingSeason]{phases}}, \code{\link[growingSeason]{seasons}}, \code{\link[growingSeason]{date_seasons}}
}
\author{
Issoufou Liman
}
