% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phases.R
\name{phases}
\alias{phases}
\alias{seasons}
\title{A constructor function for class phase}
\usage{
phases(x, type = c("v_points", "peaks"), n_criticals = 1, steps = 2,
  ts_freq = 23, returned = c("ts_seasonal", "original"))

seasons(x)
}
\arguments{
\item{x}{a numeric vector which names are dates at which the data were acquired.}

\item{type}{character string. either \code{v_points} or \code{peaks} depending on how \code{x} should be
broken into phases. \code{v_points} (default) may make more sens for expression seasonal data.}

\item{n_criticals}{Numeric. The number of points to be considered. Default to 1.}

\item{steps}{Integer. The minimum number of points at either sides of each v-like point
to be considered higher for it to qualify as v-like point. Default to 3}

\item{ts_freq}{The frequence of time series (see frequency argument in \code{\link[stats]{ts}}). Default to 23.}

\item{returned}{character string either \code{ts_seasonal} or \code{original.} if \code{ts_seasonal} (default),
the returned phases data will be the seasonal component of the decomposed time series (see \code{\link[stats]{decompose}}).
if 'original', the original data will be returned.}
}
\value{
an object of class phases comprise a list of:
\itemize{
\item{"phases"}{ list of phases}
\item{"criticals"}{index of the points separating the phases}
}
}
\description{
Takes vector (typically seasonal data such as rainfall time series) makes phases or seasons,
and returns an object of class phases or seasons. phases, are here defined
as a cycle of some sort that can be distinguished in the time series.
The only important difference between phases and seasons is that seasons are phases having a peak.
}
\examples{
## 2 years of a uni-modal pixel with complete without NAs
nam<- seq.Date(from = as.Date('2016-01-01'), to = as.Date ('2018-12-31'), by = 16)
dx11 <- c(1.30, 1.15,  1.50,  2.00,  2.01,  3.00, 3.20,  4.76,  3.50,  3.00,  2.40,  2.00,  1.50)
dx12 <-c(1.29, 1.1, 1.49, 1.99, 2, 3.1, 4.5, 4, 2.8, 2.5, 2.3, 1.6, 1.59)
dx1 <- c(dx11, dx12)
names(dx1) <- nam[1:length(dx1)]

## plotting the data
default_par <- par()
layout(rbind(c(1, 1), c(2, 3)))
par(mar = c(2, 2, 1, 1))
plot(dx1, type = 'o', main = 'raw data') # note 2 phases = 2 years and the starting point.

## returning the phases as seasonal data.
y1 <- phases(dx1, ts_freq = 12)
y1
lapply (X = y1$phases, FUN = plot, type = 'o',
 main = 'phases extracted as a seasoanal component of ts object')

## returning the phases as original data
y2 <- phases(dx1, ts_freq = 12, returned = 'original')
y2
lapply (X = y2$phases, FUN = plot, type = 'o',
main = 'phases extracted as raw')
par(default_par)

## 2 years of uni-modal pixel with many Nas towards the end.
dx2 <- dx1
dx2[21:length(dx2)] <- NA
default_par <- par()
layout(rbind(c(1, 1), c(2, 3)))
par(mar = c(2, 2, 1, 1))
plot(dx2, type = 'o')
## returning the phases as seasonal data.
y1 <- phases(dx2, ts_freq = 12)
y1
lapply (X = y1$phases, FUN = plot, type = 'o')

## returning the phases as original data
y2 <- phases(dx2, ts_freq = 12, returned = 'original')
y2
lapply (X = y2$phases, FUN = plot, type = 'o') # note 1 phase and the starting point
par(default_par)
}
\seealso{
\code{\link[SpatialProbs]{get_falls}}, \code{\link[SpatialProbs]{get_phases}}.
}
\author{
Issoufou Liman
}
