% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_ranges.R
\name{seasonal_ranges}
\alias{seasonal_ranges}
\title{Benchmark seasonal data by ranges based boxplots statistics and outliers}
\usage{
seasonal_ranges(x, size = 1000, sample = FALSE, melt = TRUE)
}
\arguments{
\item{x}{an object of class dataframe, matrix, or RasterStack. Names of x should
be dates in the format \code{X\%Y.\%m.\%d}.}

\item{size}{Integer. Sample size (if sample is TRUE or if the data cannot
be process in memory).}

\item{sample}{Logical. Should the data be sampled? If TRUE, the result is
based on a ramdon sampling of the original data.}

\item{melt}{Logical. Should the data be melted to long format? If False, the result is a list.}
}
\value{
data.frame or list (if melt is false).
}
\description{
Benchmark time series data based ranges following boxplot statistics and outliers.
}
\details{
To be documented
}
\author{
Issoufou Liman
}
